Alterei o parser+ conversão de IDs para long ints em utils.c

+++ trabalho-pratico/src/entidades/music.h
@@ -8,21 +8,21 @@
 #include <errno.h>
 #include <stdlib.h>
 #include <time.h>
-#include "utils.h"
+#include "../utils.h"
 
 // Forma de enunciar a entidade música.
 typedef struct music Music;
 
 // Função para criar uma estrutura da entidade música parametrizada.
-Music* createMusic(int id, char* title, int* artist_id, int num_artists, char* duration, char* genre, int year, char* lyrics);
+Music* createMusic(long int id, char* title, long int* artist_id, int num_artists, char* duration, char* genre, int year, char* lyrics);
 
 // Função de verificação da validade de uma musica
 int musicLineVerify(char *line);
 
 // GETTERS
-int* getMusicID(Music* m);
+long int* getMusicID(Music* m);
 char* getMusicTitle(Music* m);
-int* getMusicArtistIDs(Music* m);
+long int* getMusicArtistIDs(Music* m);
 int getMusicNumArtists(Music* m);
 char* getMusicDuration(Music* m);
 char* getMusicGenre(Music* m);
@@ -32,7 +32,7 @@ char* getMusicLyrics(Music* m);
 // SETTERS
 void setMusicID(Music* m, int newID);
 void setMusicTitle(Music* m, char* newTitle);
-void setMusicArtistIDs(Music* m, int* newArtistID, int newNumArtists);
+void setMusicArtistIDs(Music* m, long int* newArtistID, int newNumArtists);
 void setMusicDuration(Music* m, char* newDuration);
 void setMusicGenre(Music* m, char* newGenre);
 void setMusicYear(Music* m, int newYear);
